Add generic error handlers in one place, for example 403
